Rename config to runtime-config
